Tests/LibWeb/Text/expected/UIEvents/text-control-input-event-ordering.txt

input
beforeinput(value="")
textInput(value="")
after sendText(value="i")
after controlled render(value="")
textarea
beforeinput(value="")
textInput(value="")
after sendText(value="t")
after controlled render(value="")
